Changes Juju charm publish to on push
Currently, the Juju charm publish job is triggered when a PR merges,
but it will run on the merging branch instead of the main branch, causing
it to fail it the merging branch does not have the necessary credentials.

When a Pull Request merges, it results in a push, triggering
the job.