Adds a test for upgrading the charm
Currently, this charm only becomes active when the related mongodb-k8s charm
writes valid data into the relation data. This means that in any other case,
the charm will preserve whatever status is set.

Because of this, the charm remains in a Waiting state after a charm upgrade,
even though the charm is working perfectly fine.